Rework XP per user caching system
CachedXP is removed in favor of a JSON field on the user model, which can
be used to add all kinds of different caches. Caching is implemented in
this field for languages, machines and dates. This removes the need to hit
the database for other uses than to update the cache and to load the last
12 hours' XP data. As a result the profile view's display should be much
faster.